+++ lib/broadcast.py
@@ -28,7 +28,7 @@ from fractions import Fraction
 import logging
 
 from . import (util, exceptions, config, bitcoin)
-from . import bet as _bet
+from . import (bet)
 
 FORMAT = '>IdI52p'
 LENGTH = 4 + 8 + 4 + 52
@@ -122,18 +122,18 @@ def parse (db, tx, message):
     if value < 0 or value == None:
         # Cancel Open Bets?
         if value == -2:
-            cursor.execute('''SELECT * FROM bet \
+            cursor.execute('''SELECT * FROM bets \
                               WHERE (status = ? AND feed_address = ?)''',
                            ('open', tx['source']))
-            for bet in list(cursor):
-                _bet.cancel_bet(db, bet, 'dropped', tx['block_index'])
+            for i in list(cursor):
+                bet.cancel_bet(db, i, 'dropped', tx['block_index'])
         # Cancel Pending Bet Matches?
         if value == -3:
             cursor.execute('''SELECT * FROM bet_matches \
                               WHERE (status = ? AND feed_address = ?)''',
                            ('pending', tx['source']))
             for bet_match in list(cursor):
-                _bet.cancel_bet_match(db, bet_match, 'dropped', tx['block_index'])
+                bet.cancel_bet_match(db, bet_match, 'dropped', tx['block_index'])
         cursor.close()
         return